# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MarketCoin {
  id: string;
  symbol: string;
  name: string;
  image: string;
  current_price: number;
  /** @nullable */
  price_change_percentage_24h: number | null;
  market_cap: number;
  total_volume: number;
}

export type GlobalMarketResponseDataTotalMarketCap = {
  usd: number;
};

export type GlobalMarketResponseDataTotalVolume = {
  usd: number;
};

export type GlobalMarketResponseDataMarketCapPercentage = {
  btc: number;
};

export type GlobalMarketResponseData = {
  total_market_cap: GlobalMarketResponseDataTotalMarketCap;
  total_volume: GlobalMarketResponseDataTotalVolume;
  market_cap_percentage: GlobalMarketResponseDataMarketCapPercentage;
};

export interface GlobalMarketResponse {
  data: GlobalMarketResponseData;
}

export interface CoinChartResponse {
  /**
     * @items.minItems 2
     * @items.maxItems 2
     */
  prices: [number, number][];
}