# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /market/coins/{id}/chart:
    get:
      operationId: getCoinChart
      tags: [market]
      summary: Get seven-day chart data for a coin
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Seven-day market chart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoinChartResponse"
components:
  schemas:
...
    CoinChartResponse:
      type: object
      properties:
        prices:
          type: array
          items:
            type: array
            prefixItems:
              - type: number
              - type: number
            minItems: 2
            maxItems: 2
      required: [prices]