# lib/api-zod/src/generated/types/coinChartResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CoinChartResponse {
  /**
     * @items.minItems 2
     * @items.maxItems 2
     */
  prices: [number, number][];
}